[update] graphqlのオブジェクトを返すように修正した
